move `Cancorr` from manova.py into cancorr.py So is test_cancorr
